Split tests file into multiple files
